+++ r-keyvault.tf
@@ -12,7 +12,8 @@ resource "azurerm_key_vault" "keyvault" {
   enabled_for_disk_encryption     = var.enabled_for_disk_encryption
   enabled_for_template_deployment = var.enabled_for_template_deployment
 
-  purge_protection_enabled = var.purge_protection_enabled
+  purge_protection_enabled   = var.purge_protection_enabled
+  soft_delete_retention_days = var.soft_delete_retention_days
 
   enable_rbac_authorization = var.rbac_authorization_enabled
 
@@ -21,8 +22,8 @@ resource "azurerm_key_vault" "keyvault" {
     iterator = acl
 
     content {
-      bypass                     = coalesce(acl.value.bypass, "None")
-      default_action             = coalesce(acl.value.default_action, "Deny")
+      bypass                     = acl.value.bypass
+      default_action             = acl.value.default_action
       ip_rules                   = acl.value.ip_rules
       virtual_network_subnet_ids = acl.value.virtual_network_subnet_ids
     }